feat: privateRouter 기존 App 에서 분리 후 적용
